Descriptive statistics for the clusters identified by the Poisson kernel-based clustering.

Method for objects of class pkbc which computes some descriptive for each variable with respect to the detected groups.

Method for objects of class pkbc which computes some descriptive for each variable with respect to the detected groups.

Usage

extract_stats(object, ...)

# S4 method for pkbc
extract_stats(object, k)

Arguments

object

Object of class pkbc.

...

possible additional inputs

k

Number of clusters to be used.

Value

List with computed descriptive statistics for each variable.

Details

The function computes mean, standard deviation, median, inter-quantile range, minimum and maximum for each variable in the data set given the final membership assigned by the clustering algorithm.

Examples

#We generate three samples of 100 observations from 3-dimensional
#Poisson kernel-based densities with rho=0.8 and different mean directions
dat<-matrix(rnorm(300),ncol=3)

#Perform the clustering algorithm
pkbc_res<- pkbc(dat, 3)
extract_stats(pkbc_res, 3)
#> [[1]]
#>           Group 1       Group 2    Group 3     Overall
#> mean    0.1751353 -0.4322713256 -0.7386501  0.05333293
#> sd      0.5524869  0.0007700714  0.1577503  0.59876670
#> median  0.1731522 -0.4322713256 -0.7151248  0.02577322
#> IQR     0.8802893  0.0005445227  0.2028542  1.07860438
#> min    -0.9051036 -0.4328158484 -0.9673199 -0.96731993
#> max     0.9953733 -0.4317268029 -0.4091660  0.99537326
#> 
#> [[2]]
#>            Group 1    Group 2    Group 3     Overall
#> mean    0.04249666 0.51297091 -0.5500070 -0.01919429
#> sd      0.55921676 0.01916986  0.2227660  0.56320304
#> median  0.06528290 0.51297091 -0.5957768 -0.02631685
#> IQR     0.87000688 0.01355514  0.3479383  0.97848746
#> min    -0.99573191 0.49941577 -0.8732286 -0.99573191
#> max     0.99576330 0.52652605 -0.1931502  0.99576330
#> 
#> [[3]]
#>           Group 1      Group 2    Group 3    Overall
#> mean    0.1425588 -0.741441983  0.1054550  0.1204263
#> sd      0.5834184  0.012813818  0.2811668  0.5625722
#> median  0.2285537 -0.741441983  0.1372116  0.1789860
#> IQR     0.9615483  0.009060738  0.3522091  0.8972161
#> min    -0.9924515 -0.750502720 -0.4418849 -0.9924515
#> max     0.9949503 -0.732381245  0.4348400  0.9949503
#>