An S4 class for kernel-based distance tests with normal kernel

A class to represent the results of Gaussian kernel-based quadratic distance tests. This includes the normality test, the two-sample test statistics and the k-sample tests.

Slots

method

String indicating the kernel-based quadratic distance test performed.

Un

The value of the test U-statistics.

Vn

The value of the test V-statistic.

H0_Un

A logical value indicating whether or not the null hypothesis is rejected according to U-statistics.

H0_Vn

A logical value indicating whether or not the null hypothesis is rejected according to Vn.

data

List of samples X (and Y).

CV_Un

The critical value computed for the test Un.

CV_Vn

The critical value computed for the test Vn.

cv_method

The method used to estimate the critical value (one of "subsampling", "permutation" or "bootstrap").

h

A list with the value of bandwidth parameter used for the Gaussian kernel. If the function select_h is used, then also the matrix of computed power values and the resulting power plot are provided.

B

Number of bootstrap/permutation/subsampling replications.

var_Un

exact variance of the kernel-based U-statistic.

See also

Examples

# create a kb.test object
x <- matrix(rnorm(100),ncol=2)
y <- matrix(rnorm(100),ncol=2)
# Normality test
kb.test(x, h=0.5)
#> 
#>  Kernel-based quadratic distance Normality test 
#> 		U-statistic	V-statistic
#> ------------------------------------------------
#> Test Statistic:	 1.310854 	 0.7979308 
#> Critical Value:	 2.4409 	 6.071062 
#> H0 is rejected:	 FALSE 		 FALSE 
#> Selected tuning parameter h:  0.5 
#> 

# Two-sample test
kb.test(x,y,h=0.5, method="subsampling",b=0.9)
#> 
#>  Kernel-based quadratic distance two-sample test 
#> U-statistics	 Dn 		 Trace 
#> ------------------------------------------------
#> Test Statistic:	 0.1153108 	 0.133021 
#> Critical Value:	 1.407721 	 1.625753 
#> H0 is rejected:	 FALSE 		 FALSE 
#> CV method:  subsampling 
#> Selected tuning parameter h:  0.5 
#>