An S4 class for Poisson kernel-based quadratic distance tests.

A class to represent the results of Poisson kernel-based quadratic distance tests for Uniformity on the sphere.

Slots

method

The method used for the test ("Poisson Kernel-based quadratic distance test of Uniformity on the Sphere").

x

Matrix of data

Un

The value of the U-statistic.

CV_Un

The critical value for Un computed through replications.

H0_Un

A logical value indicating whether or not the null hypothesis is rejected according to Un.

Vn

The value of the V-statistic.

CV_Vn

The critical value for Vn computed following the asymptotic distribution.

H0_Vn

A logical value indicating whether or not the null hypothesis is rejected according to Vn.

rho

The concentration parameter of the Poisson kernel.

B

Number of replications.

var_Un

exact variance of the kernel-based U-statistic.

Examples

# create a pk.test object
d=3
size=200
x_sp <- sample_hypersphere(d, n_points=size)
pk.test(x_sp,rho=0.8)
#> 
#>  Poisson Kernel-based quadratic distance test of 
#>                         Uniformity on the Sphere 
#> Selected consentration parameter rho:  0.8 
#> 
#> U-statistic:
#> 
#> H0 is rejected:  FALSE 
#> Statistic Un:  -0.1438295 
#> Critical value:  1.52535 
#> 
#> V-statistic:
#> 
#> H0 is rejected:  FALSE 
#> Statistic Vn:  43.30734 
#> Critical value:  52.23077 
#>