Plotting method for Poisson kernel-based clustering

Plots for a pkbc object.

Usage

# S4 method for class 'pkbc,ANY'
plot(x, true_label = NULL, pca_res = FALSE)

Arguments

x

Object of class pkbc

true_label

factor or vector of true membership to clusters (if available). It must have the same length of final memberships.

pca_res

Logical. If TRUE the results from PCALocantore when dimension is greater than 3 are also reported.

Value

One of the following plot:

scatterplot of data points colored by final membership

elbow plot

Details

scatterplot: If dimension is equal to 2 or 3, points are displayed on the circle and sphere, respectively. If dimension if greater than 3, the spherical Principal Component procedure proposed by Locantore et al., (1999) is applied for dimensionality reduction and the first three principal components are normalized and displayed on the sphere. For d > 3, the complete results from the PcaLocantore function (package rrcov) are returned if pca_res=TRUE.

elbow plot: the within cluster sum of squares (wcss) is computed using the Euclidean distance and the cosine similarity.

References

Locantore, N., Marron, J.S., Simpson, D.G. et al. (1999) "Robust principal component analysis for functional data." Test 8, 1–73. https://doi.org/10.1007/BF02595862

Examples

# \donttest{
dat<-matrix(rnorm(300),ncol=3)
pkbc_res<- pkbc(dat, 3)
stats_clusters(pkbc_res, 3)
#> [[1]]
#>            Group 1    Group 2   Group 3       Overall
#> mean   -0.11423666  0.3590187 0.5147192  0.0009016251
#> sd      0.59540850  0.3076964        NA  0.5764088492
#> median -0.08317254  0.3416886 0.5147192  0.0529415862
#> IQR     1.08582765  0.4552225 0.0000000  0.9657622803
#> min    -0.99815567 -0.2498777 0.5147192 -0.9981556667
#> max     0.92996966  0.8834617 0.5147192  0.9299696565
#> 
#> [[2]]
#>           Group 1     Group 2    Group 3     Overall
#> mean    0.0746302  0.02798259 -0.6126422  0.05702852
#> sd      0.5898005  0.44361604         NA  0.55877530
#> median  0.1263663  0.04841560 -0.6126422  0.07707060
#> IQR     0.9728322  0.67595429  0.0000000  0.86676405
#> min    -0.9993915 -0.71110682 -0.6126422 -0.99939152
#> max     0.9987495  0.75635249 -0.6126422  0.99874954
#> 
#> [[3]]
#>           Group 1    Group 2   Group 3      Overall
#> mean    0.1984374 -0.7452149 -0.599778 -0.026584810
#> sd      0.5021823  0.1944794        NA  0.601351973
#> median  0.1485244 -0.7525265 -0.599778 -0.006964457
#> IQR     0.7709032  0.3395141  0.000000  1.063345824
#> min    -0.7126606 -0.9970685 -0.599778 -0.997068462
#> max     0.9813179 -0.3946371 -0.599778  0.981317913
#> 
# }