Cluster spherical observations using a mixture of Poisson kernel-based densities

Obtain predictions of membership for spherical observations based on a mixture of Poisson kernel-based densities estimated by pkbc

# S4 method for class 'pkbc'
predict(object, k, newdata = NULL)

Arguments

object

Object of class pkbc

k

Number of clusters to be used.

newdata

a data.frame or a matrix of the data. If missing the clustering data obtained from the pkbc object are classified.

Value

Returns a list with the following components

Memb: vector of predicted memberships of newdata

Probs: matrix where entry (i,j) denotes the probability that observation i belongs to the k-th cluster.

See also

Examples

# generate data
dat <- rbind(matrix(rnorm(100), ncol = 2), matrix(rnorm(100, 5), ncol = 2))
res <- pkbc(dat, 2)

# extract membership of dat
predict(res, k = 2)
#>   [1] 1 2 1 2 2 2 2 2 2 2 2 2 2 2 2 2 1 2 2 2 2 2 2 2 2 1 1 1 2 1 2 2 2 2 2 2 2
#>  [38] 1 2 1 1 2 2 2 1 1 1 2 2 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
# predict membership of new data
newdat <- rbind(matrix(rnorm(10), ncol = 2), matrix(rnorm(10, 5), ncol = 2))
predict(res, k = 2, newdat)
#> $Memb
#>  [1] 1 2 2 1 2 1 1 1 1 1
#> 
#> $Probs
#>             [,1]       [,2]
#>  [1,] 0.92527292 0.07472708
#>  [2,] 0.05863774 0.94136226
#>  [3,] 0.11560241 0.88439759
#>  [4,] 0.65760025 0.34239975
#>  [5,] 0.08235908 0.91764092
#>  [6,] 0.96052592 0.03947408
#>  [7,] 0.97923782 0.02076218
#>  [8,] 0.93746395 0.06253605
#>  [9,] 0.97964713 0.02035287
#> [10,] 0.97033286 0.02966714
#>