Descriptive statistics for the clusters identified by the Poisson kernel-based clustering.

Method for objects of class pkbc which computes some descriptive for each variable with respect to the detected groups.

Method for objects of class pkbc which computes descriptive statistics for each variable with respect to the detected groups.

stats_clusters(object, ...)

# S4 method for class 'pkbc'
stats_clusters(object, k)

Arguments

object

Object of class pkbc.

...

possible additional inputs

k

Number of clusters to be used.

Value

List with computed descriptive statistics for each dimension.

Details

The function computes mean, standard deviation, median, inter-quantile range, minimum and maximum for each variable in the data set given the final membership assigned by the clustering algorithm.

See also

Examples

#We generate three samples of 100 observations from 3-dimensional
#Poisson kernel-based densities with rho=0.8 and different mean directions
dat<-matrix(rnorm(300),ncol=3)

#Perform the clustering algorithm
pkbc_res<- pkbc(dat, 3)
stats_clusters(pkbc_res, 3)
#> [[1]]
#>            Group 1   Group 2     Group 3     Overall
#> mean   -0.07250026 0.8770771  0.02418397 -0.05913712
#> sd      0.55954051        NA  0.10119157  0.55397548
#> median -0.14806589 0.8770771  0.04822803 -0.11021321
#> IQR     0.82595800 0.0000000  0.12929025  0.81538105
#> min    -0.99439702 0.8770771 -0.10796947 -0.99439702
#> max     0.94994816 0.8770771  0.10824930  0.94994816
#> 
#> [[2]]
#>            Group 1   Group 2     Group 3    Overall
#> mean   -0.09236276 0.3804107 -0.41253893 -0.1004421
#> sd      0.58536436        NA  0.06043979  0.5760124
#> median -0.14452463 0.3804107 -0.42344668 -0.1640092
#> IQR     0.99407299 0.0000000  0.08874042  0.9781217
#> min    -0.99349220 0.3804107 -0.46276582 -0.9934922
#> max     0.96437223 0.3804107 -0.34049652  0.9643722
#> 
#> [[3]]
#>            Group 1    Group 2    Group 3     Overall
#> mean    0.02400049 -0.2932975 0.90460289  0.05605160
#> sd      0.58342548         NA 0.02584192  0.59542035
#> median  0.02864568 -0.2932975 0.90159928  0.05619702
#> IQR     0.98410862  0.0000000 0.03642008  0.98250679
#> min    -0.96992620 -0.2932975 0.88012879 -0.96992620
#> max     0.99420069 -0.2932975 0.93508422  0.99420069
#>