Descriptive statistics for Poisson kernel-based clustering results.

Method for objects of class pkbc which computes and displays some descriptive for each variable with respect to the detected groups. It also generates a plot of data points colored by the final membership.

Usage

summary_stat(object, k, true_label = NULL)

Arguments

object

Object of class pkbc.

k

Number of clusters to be used.

true_label

Vector of true memberships to clusters (default: NULL).

Value

List with computed descriptive statistics for each variable. For d > 3, the complete results from the PcaLocantore function (package rrcov) are also returned.

Details

The function computes mean, standard deviation, median, inter-quantile range, minimum and maximum for each variable in the data set given the final membership assigned by the clustering algorithm. If dimension is equal to 2 or 3, points are displayed on the circle and sphere, respectively. If dimension if greater than 3, the spherical Principal Component procedure proposed by Locantore et al., (1999) is applied for dimensionality reduction and the first three principal components are normalized and displayed on the sphere.

References

Locantore, N., Marron, J.S., Simpson, D.G. et al. (1999) "Robust principal component analysis for functional data." Test 8, 1–73. https://doi.org/10.1007/BF02595862

Examples

#We generate three samples of 100 observations from 3-dimensional
#Poisson kernel-based densities with rho=0.8 and different mean directions
size<-100
groups<-c(rep(1, size), rep(2, size),rep(3,size))
rho<-0.8
set.seed(081423)
data1<-rpkb(size, c(1,0,0),rho,method='rejvmf')
data2<-rpkb(size, c(0,1,0),rho,method='rejvmf')
data3<-rpkb(size, c(-1,0,0),rho,method='rejvmf')
data<-rbind(data1$x,data2$x, data3$x)

#Perform the clustering algorithm
pkbc_res<- pkbc(data, 3)
summary_stat(pkbc_res, 3)
#> $metrics
#> $metrics[[1]]
#>          Group 1     Group 2     Group 3      Overall
#> mean   0.8715954  0.01868522 -0.85332703 -0.005122387
#> sd     0.1625037  0.25354878  0.20116183  0.745152609
#> median 0.9328949  0.02007526 -0.92187452  0.009839853
#> IQR    0.1155861  0.32644557  0.15273214  1.726651633
#> min    0.3106241 -0.62856435 -0.99986735 -0.999867352
#> max    0.9999498  0.63056946 -0.09929114  0.999949817
#> 
#> $metrics[[2]]
#>             Group 1   Group 2      Group 3    Overall
#> mean   -0.040570128 0.9033756 -0.005035596  0.2677551
#> sd      0.305064729 0.1188987  0.315272544  0.5052780
#> median  0.003805487 0.9594194  0.005197609  0.1310688
#> IQR     0.342097544 0.1274881  0.328562370  0.9307831
#> min    -0.939376625 0.3970333 -0.800340937 -0.9393766
#> max     0.601805719 0.9998137  0.714705460  0.9998137
#> 
#> $metrics[[3]]
#>             Group 1     Group 2     Group 3     Overall
#> mean    0.014712275 -0.04491125  0.06288442  0.01305106
#> sd      0.348419022  0.32415217  0.36130499  0.34735638
#> median -0.001347224  0.01105224  0.04476277  0.01664902
#> IQR     0.432630201  0.32888851  0.45513734  0.41340433
#> min    -0.784086368 -0.89290846 -0.96838195 -0.96838195
#> max     0.830930688  0.81869150  0.96703790  0.96703790
#> 
#>