Breast Cancer Wisconsin (Diagnostic)

The breast_cancer Wisconsin data has 569 rows and 31 columns. The first 30 variables report the features that are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass. They describe characteristics of the cell nuclei present in the image. The last column indicates the class labels (Benign = 0 or Malignant = 1).

Usage

breast_cancer

Format

A data frame of 569 observations and 31 variables.

Source

Wolberg,William, Mangasarian,Olvi, Street,Nick, and Street,W.. (1995). Breast Cancer Wisconsin (Diagnostic). UCI Machine Learning Repository.

https://doi.org/10.24432/C5DW2B.

References

Street, W. N., Wolberg, W. H., & Mangasarian, O. L. (1993, July). Nuclear feature extraction for breast tumor diagnosis. In Biomedical image processing and biomedical visualization (Vol. 1905, pp. 861-870). SPIE.

Examples

data(breast_cancer)
summary(breast_cancer)
#>     radius1          texture1       perimeter1         area1       
#>  Min.   : 6.981   Min.   : 9.71   Min.   : 43.79   Min.   : 143.5  
#>  1st Qu.:11.700   1st Qu.:16.17   1st Qu.: 75.17   1st Qu.: 420.3  
#>  Median :13.370   Median :18.84   Median : 86.24   Median : 551.1  
#>  Mean   :14.127   Mean   :19.29   Mean   : 91.97   Mean   : 654.9  
#>  3rd Qu.:15.780   3rd Qu.:21.80   3rd Qu.:104.10   3rd Qu.: 782.7  
#>  Max.   :28.110   Max.   :39.28   Max.   :188.50   Max.   :2501.0  
#>   smoothness1       compactness1       concavity1      concave_points1  
#>  Min.   :0.05263   Min.   :0.01938   Min.   :0.00000   Min.   :0.00000  
#>  1st Qu.:0.08637   1st Qu.:0.06492   1st Qu.:0.02956   1st Qu.:0.02031  
#>  Median :0.09587   Median :0.09263   Median :0.06154   Median :0.03350  
#>  Mean   :0.09636   Mean   :0.10434   Mean   :0.08880   Mean   :0.04892  
#>  3rd Qu.:0.10530   3rd Qu.:0.13040   3rd Qu.:0.13070   3rd Qu.:0.07400  
#>  Max.   :0.16340   Max.   :0.34540   Max.   :0.42680   Max.   :0.20120  
#>    symmetry1      fractal_dimension1    radius2          texture2     
#>  Min.   :0.1060   Min.   :0.04996    Min.   :0.1115   Min.   :0.3602  
#>  1st Qu.:0.1619   1st Qu.:0.05770    1st Qu.:0.2324   1st Qu.:0.8339  
#>  Median :0.1792   Median :0.06154    Median :0.3242   Median :1.1080  
#>  Mean   :0.1812   Mean   :0.06280    Mean   :0.4052   Mean   :1.2169  
#>  3rd Qu.:0.1957   3rd Qu.:0.06612    3rd Qu.:0.4789   3rd Qu.:1.4740  
#>  Max.   :0.3040   Max.   :0.09744    Max.   :2.8730   Max.   :4.8850  
#>    perimeter2         area2          smoothness2        compactness2     
#>  Min.   : 0.757   Min.   :  6.802   Min.   :0.001713   Min.   :0.002252  
#>  1st Qu.: 1.606   1st Qu.: 17.850   1st Qu.:0.005169   1st Qu.:0.013080  
#>  Median : 2.287   Median : 24.530   Median :0.006380   Median :0.020450  
#>  Mean   : 2.866   Mean   : 40.337   Mean   :0.007041   Mean   :0.025478  
#>  3rd Qu.: 3.357   3rd Qu.: 45.190   3rd Qu.:0.008146   3rd Qu.:0.032450  
#>  Max.   :21.980   Max.   :542.200   Max.   :0.031130   Max.   :0.135400  
#>    concavity2      concave_points2      symmetry2        fractal_dimension2 
#>  Min.   :0.00000   Min.   :0.000000   Min.   :0.007882   Min.   :0.0008948  
#>  1st Qu.:0.01509   1st Qu.:0.007638   1st Qu.:0.015160   1st Qu.:0.0022480  
#>  Median :0.02589   Median :0.010930   Median :0.018730   Median :0.0031870  
#>  Mean   :0.03189   Mean   :0.011796   Mean   :0.020542   Mean   :0.0037949  
#>  3rd Qu.:0.04205   3rd Qu.:0.014710   3rd Qu.:0.023480   3rd Qu.:0.0045580  
#>  Max.   :0.39600   Max.   :0.052790   Max.   :0.078950   Max.   :0.0298400  
#>     radius3         texture3       perimeter3         area3       
#>  Min.   : 7.93   Min.   :12.02   Min.   : 50.41   Min.   : 185.2  
#>  1st Qu.:13.01   1st Qu.:21.08   1st Qu.: 84.11   1st Qu.: 515.3  
#>  Median :14.97   Median :25.41   Median : 97.66   Median : 686.5  
#>  Mean   :16.27   Mean   :25.68   Mean   :107.26   Mean   : 880.6  
#>  3rd Qu.:18.79   3rd Qu.:29.72   3rd Qu.:125.40   3rd Qu.:1084.0  
#>  Max.   :36.04   Max.   :49.54   Max.   :251.20   Max.   :4254.0  
#>   smoothness3       compactness3       concavity3     concave_points3  
#>  Min.   :0.07117   Min.   :0.02729   Min.   :0.0000   Min.   :0.00000  
#>  1st Qu.:0.11660   1st Qu.:0.14720   1st Qu.:0.1145   1st Qu.:0.06493  
#>  Median :0.13130   Median :0.21190   Median :0.2267   Median :0.09993  
#>  Mean   :0.13237   Mean   :0.25427   Mean   :0.2722   Mean   :0.11461  
#>  3rd Qu.:0.14600   3rd Qu.:0.33910   3rd Qu.:0.3829   3rd Qu.:0.16140  
#>  Max.   :0.22260   Max.   :1.05800   Max.   :1.2520   Max.   :0.29100  
#>    symmetry3      fractal_dimension3      y            
#>  Min.   :0.1565   Min.   :0.05504    Length:569        
#>  1st Qu.:0.2504   1st Qu.:0.07146    Class :character  
#>  Median :0.2822   Median :0.08004    Mode  :character  
#>  Mean   :0.2901   Mean   :0.08395                      
#>  3rd Qu.:0.3179   3rd Qu.:0.09208                      
#>  Max.   :0.6638   Max.   :0.20750