Plotting method for Poisson kernel-based clustering

Plots for a pkbc object.

Usage

# S4 method for class 'pkbc,ANY'
plot(x, k = NULL, true_label = NULL, pca_res = FALSE)

Arguments

x

Object of class pkbc

k

number of considered clusters. If it is not provided the scatter plot is displayed for each value of number of clusters present in the x object

true_label

factor or vector of true membership to clusters (if available). It must have the same length of final memberships.

pca_res

Logical. If TRUE the results from PCALocantore are also reported (when dimension is greater than 3).

Value

scatterplot of data points colored by final membership

elbow plot

Details

scatterplot: If dimension is equal to 2 or 3, points are displayed on the circle and sphere, respectively. If dimension if greater than 3, the spherical Principal Component procedure proposed by Locantore et al. (1999), is applied for dimensionality reduction and the first three principal components are normalized and displayed on the sphere. For d > 3, the complete results from the PcaLocantore function (package rrcov) are returned if pca_res=TRUE.

elbow plot: the within cluster sum of squares (wcss) is computed using the Euclidean distance (left) and the cosine similarity (right).

References

Locantore, N., Marron, J.S., Simpson, D.G. et al. (1999) "Robust principal component analysis for functional data." Test 8, 1–73. https://doi.org/10.1007/BF02595862

See also

Examples

dat<-matrix(rnorm(300),ncol=3)
pkbc_res<- pkbc(dat, 3)
stats_clusters(pkbc_res, 3)
#> [[1]]
#>            Group 1    Group 2     Group 3     Overall
#> mean    0.08616614 -0.8522997 0.306305518  0.03032233
#> sd      0.55622334  0.1139002 0.239097130  0.57609801
#> median  0.09893314 -0.8590883 0.220771423  0.08528945
#> IQR     0.95811426  0.1154519 0.319424189  0.91819081
#> min    -0.99815567 -0.9929153 0.007812835 -0.99815567
#> max     0.97484325 -0.6046193 0.747318680  0.97484325
#> 
#> [[2]]
#>            Group 1   Group 2    Group 3      Overall
#> mean   -0.06949009 0.3329087  0.1335748 -0.006875767
#> sd      0.58955595 0.1928422  0.3058553  0.548503665
#> median -0.12515153 0.3197409  0.2780209  0.020922028
#> IQR     1.05338258 0.3475881  0.4344834  0.891655218
#> min    -0.99465804 0.1143958 -0.3618267 -0.994658036
#> max     0.99874954 0.5919510  0.5345903  0.998749545
#> 
#> [[3]]
#>           Group 1    Group 2    Group 3     Overall
#> mean   -0.2862676 0.28557672 0.86077538 -0.08568599
#> sd      0.5104903 0.20295785 0.09470435  0.60742332
#> median -0.3332680 0.29047540 0.88064534 -0.07405065
#> IQR     0.7500978 0.33726926 0.12384691  1.06925832
#> min    -0.9996330 0.01117247 0.66139978 -0.99963299
#> max     0.8079944 0.53294413 0.98131791  0.98131791
#>