Plotting method for Poisson kernel-based clustering

Plots for a pkbc object.

Usage

# S4 method for pkbc,ANY
plot(x, y = NULL, true_label = NULL, pca_res = FALSE)

Arguments

x

Object of class pkbc

y

formal input

true_label

factor or vector of true membership to clusters (if available). It must have the same length of final memberships.

pca_res

Logical. If TRUE the results from PCALocantore when dimension is greater than 3 are also reported.

Value

One of the following plot:

scatterplot of data points colored by final membership

elbow plot

Details

scatterplot: If dimension is equal to 2 or 3, points are displayed on the circle and sphere, respectively. If dimension if greater than 3, the spherical Principal Component procedure proposed by Locantore et al., (1999) is applied for dimensionality reduction and the first three principal components are normalized and displayed on the sphere. For d > 3, the ù complete results from the PcaLocantore function (package rrcov) are returned if pca_res=TRUE.

elbow plot: the within cluster sum of squares (wcss) is computed using the Euclidean distance and the cosine similarity.

References

Locantore, N., Marron, J.S., Simpson, D.G. et al. (1999) "Robust principal component analysis for functional data." Test 8, 1–73. https://doi.org/10.1007/BF02595862

Examples

# \donttest{
dat<-matrix(rnorm(300),ncol=3)
pkbc_res<- pkbc(dat, 3)
extract_stats(pkbc_res, 3)
#> [[1]]
#>           Group 1     Group 2     Group 3      Overall
#> mean   -0.3581548 -0.78495693  0.05274666  0.002961461
#> sd      0.1011887  0.04067301  0.59893653  0.596370987
#> median -0.3315073 -0.78629283  0.10214240  0.010785524
#> IQR     0.1107611  0.04065655  1.01060152  0.991783489
#> min    -0.5282233 -0.82494553 -0.97541567 -0.975415666
#> max    -0.2546980 -0.74363242  0.99207654  0.992076543
#> 
#> [[2]]
#>           Group 1     Group 2      Group 3     Overall
#> mean   0.86871524 -0.52259544  0.009118631  0.04474301
#> sd     0.08048405  0.04384318  0.578106548  0.59687373
#> median 0.87833762 -0.50530082 -0.028918855  0.01162669
#> IQR    0.10139593  0.04120554  1.033174684  1.13887446
#> min    0.74021494 -0.57244828 -0.957970785 -0.95797078
#> max    0.95153787 -0.49003721  0.980640500  0.98064050
#> 
#> [[3]]
#>           Group 1    Group 2     Group 3     Overall
#> mean   -0.2944026 -0.3275399  0.01338372 -0.01531117
#> sd      0.1405221  0.0400607  0.56133827  0.54398216
#> median -0.3260401 -0.3454183  0.01488409 -0.09418980
#> IQR     0.2197151  0.0369477  0.94827983  0.86470208
#> min    -0.4496203 -0.3555484 -0.98406028 -0.98406028
#> max    -0.1237515 -0.2816530  0.98548914  0.98548914
#> 
# }